Shifted functions from routes to controllers
