+++ backend/src/controllers/userController.js
@@ -45,23 +45,6 @@ const registerUser = asyncHandler(async (req, res) => {
   }
 });
 
-// const users = asyncHandler(async (req, res) => {
-//   const filter = req.query.name;
-//   const filteredUser = await prisma.user.findMany({
-//     where: {
-//       OR: [
-//         {
-//           username: {
-//             contains: filter,
-//             mode: "insensitive", // Optional: makes the search case-insensitive
-//           },
-//         },
-//       ],
-//     },
-//   });
-//   res.json(filteredUser);
-// });
-
 const loginUser = asyncHandler(async (req, res) => {
   const { email, password } = req.body;
   let user = await prisma.User.findUnique({
@@ -103,4 +86,99 @@ const loginUser = asyncHandler(async (req, res) => {
   res.json(user);
 });
 
-export { registerUser, loginUser };
+const signOut = asyncHandler(async (req, res) => {
+  res.clearCookie("Authorization");
+  res.sendStatus(200);
+});
+
+const updateFields = asyncHandler(async (req, res) => {
+  try {
+    const { username, name } = req.body;
+    if (!username || !name) {
+      return res.status(422).json({ error: "Please add all the fields" });
+    }
+
+    const user = await prisma.user.update({
+      where: {
+        id: req.user.id,
+      },
+      data: {
+        name,
+        username,
+      },
+    });
+    res.json({
+      username: user.username,
+      name: user.name,
+      email: user.email,
+    });
+  } catch (e) {
+    res.status(500).json({ error: "Failed to update user" });
+  }
+});
+
+const updatepassword = asyncHandler(async (req, res) => {
+  try {
+    const { password } = req.body;
+    const hashedPassword = bcrypt.hashSync(password, 8);
+    if (!password) {
+      return res.status(422).json({ error: "Please add all the fields" });
+    }
+    const user = await prisma.user.update({
+      where: {
+        id: req.user.id,
+      },
+      data: {
+        password: hashedPassword,
+      },
+    });
+    res.json({
+      username: user.username,
+      name: user.name,
+      email: user.email,
+    });
+  } catch (e) {
+    res.status(500).json({ error: "Failed to update user" });
+  }
+});
+
+const filterUsers = asyncHandler(async (req, res) => {
+  let filter = req.query.filter || "";
+  filter = filter.replace(/^"|"$/g, "");
+
+  const users = await prisma.User.findMany({
+    where: {
+      OR: [
+        {
+          username: {
+            contains: filter,
+            mode: "insensitive",
+          },
+        },
+      ],
+    },
+    select: {
+      id: true,
+      username: true,
+      Stats: {
+        select: {
+          totalTimeStudied: true,
+          maxTimeStudied: true,
+          streak: true,
+        },
+      },
+    },
+  });
+  res.json({
+    users,
+  });
+});
+
+export {
+  registerUser,
+  loginUser,
+  signOut,
+  updateFields,
+  updatepassword,
+  filterUsers,
+};
