+++ backend/src/routes/statsRoutes.js
@@ -1,9 +1,9 @@
 import express from "express";
-import prisma from "../lib/prisma.js";
 import {
   saveStats,
   fetchStats,
   fetchHistory,
+  fetchBulkData,
 } from "../controllers/statsControllers.js";
 import checkAuth from "../middleware/checkAuth.js";
 
@@ -11,28 +11,6 @@ const router = express.Router();
 
 router.post("/", checkAuth, saveStats).get("/:id", fetchStats);
 router.get("/history/:id", fetchHistory);
-
-router.get("/bulkdata", checkAuth, async (req, res) => {
-  const { username } = req.body;
-  if (!username) {
-    return res.status(422).json({ error: "Please add all the fields" });
-  }
-  try {
-    const stats = await prisma.History.findUnique({
-      where: {
-        username,
-      },
-      select: {
-        date: true,
-        timeStudied: true,
-      },
-    });
-    res.json({
-      date: stats.date,
-      timeStudied: stats.timeStudied,
-    });
-  } catch (e) {
-    res.status(500).json({ error: "Failed to fetch stats" });
-  }
-});
+router.get("/bulkdata", checkAuth, fetchBulkData);
+//fetchBulkData is a route that will display the History with which the graph can be made
 export default router;
